Adds CallbackTrait for checking if callbacks are registered.
